Move bincode out of nym-sphinx
