+++ common/nymsphinx/src/lib.rs
@@ -22,10 +22,3 @@ pub use nym_sphinx_framing as framing;
 
 // TEMP UNTIL FURTHER REFACTORING
 pub use preparer::payload::NymPayloadBuilder;
-
-fn make_bincode_serializer() -> impl bincode::Options {
-    use bincode::Options;
-    bincode::DefaultOptions::new()
-        .with_big_endian()
-        .with_varint_encoding()
-}
